requirements.txt
# Core
opencv-python>=4.8.0
mediapipe>=0.10.0
ultralytics>=8.0.0
numpy>=1.24.0

# Configuration
pydantic>=2.0.0
pyyaml>=6.0.0

# System monitoring
psutil>=5.9.0

# CLI
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.0
